requirements-dev.txt
Hydrogram==0.1.4
TgCrypto
django
uvloop
xxhash
loguru
python-dotenv
black
apscheduler







